implemennted correct behaviour when services not configured
