+++ src/app/app.component.ts
@@ -1,3 +1,5 @@
+import { AlertService } from './../services/alert.service';
+import { PushService } from './../services/push.service';
 import { Component, ViewChild } from '@angular/core';
 import { Nav, Platform, MenuController } from 'ionic-angular';
 import { StatusBar } from '@ionic-native/status-bar';
@@ -18,6 +20,8 @@ import { PushMessagesPage } from '../pages/pushMessages/pushMessages';
 import { TrustCheckPage } from '../pages/trustCheck/trustCheck';
 import { DocumentationPage } from '../pages/documentation/documentation';
 import { DeviceSecurityPage } from '../pages/deviceSecurity/deviceSecurity';
+import { Auth } from '@aerogear/auth';
+import { constants } from '../constants/constants';
 
 @Component({
   templateUrl: 'app.html'
@@ -32,7 +36,9 @@ export class MyApp {
   constructor(public platform: Platform,
               public statusBar: StatusBar,
               public splashScreen: SplashScreen,
-              private menuCtrl: MenuController) {
+              private menuCtrl: MenuController,
+              private auth:Auth,
+              private alert: AlertService) {
     this.initializeApp();
 
   }
@@ -81,8 +87,26 @@ export class MyApp {
   openPage(page) {
     // Reset the content nav to have just this page
     // we wouldn't want the back button to show in this scenario
-    this.menuCtrl.close().then(() => {  
-      this.nav.setRoot(page.component, { 'linkParam' : page.param });
+    this.menuCtrl.close().then(() => {
+      let changePageAllowed: boolean = true;
+      if(page.component === PushMessagesPage) {
+        if(!PushService.registered) {
+          changePageAllowed = false;
+          this.alert.showAlert(constants.pushAlertMessage, constants.featureNotConfigured, 
+            constants.alertButtons, constants.showDocs, constants.pushDocsUrl);
+        } 
+      } else if (page.component === AuthPage) {
+        const configKeys: string[] = Object.keys(this.auth.getConfig()).toString().split(',');
+            console.log(configKeys.length);
+            if (configKeys.length <= 1 || configKeys.indexOf('url') === -1) {
+              changePageAllowed = false;
+              this.alert.showAlert(constants.idmMessage, constants.featureNotConfigured, 
+                constants.alertButtons, constants.showDocs, constants.idmUrl);
+            }
+      }
+      if(changePageAllowed) {
+        this.nav.setRoot(page.component, { 'linkParam' : page.param });
+      }
     })
   }
 }
