+++ src/pages/auth/auth.ts
@@ -4,8 +4,6 @@ import { authProvider } from '../../services/auth.service';
 import { Auth } from '@aerogear/auth';
 import { ToastController } from 'ionic-angular';
 import { AuthDetailsPage } from '../authDetails/authDetails';
-import { AlertService } from '../../services/alert.service';
-import { constants } from '../../constants/constants';
 
 @Component({
     selector: 'page-auth',
@@ -14,10 +12,9 @@ import { constants } from '../../constants/constants';
 })
 export class AuthPage {
     authButtonState: boolean;
-    configValid: boolean;
 
     constructor(public toastCtrl: ToastController, private auth: Auth, public navCtrl: NavController, 
-        public navParams: NavParams, private alert: AlertService) {
+        public navParams: NavParams) {
         this.auth = auth;
         this.authButtonState = true;
         this.toastCtrl = toastCtrl;
@@ -32,16 +29,6 @@ export class AuthPage {
     ionViewDidEnter(): void {
         if (this.auth.isAuthenticated()) {
             this.navCtrl.setRoot(AuthDetailsPage);
-        } else {
-            const configKeys: string[] = Object.keys(this.auth.getConfig()).toString().split(',');
-            console.log(configKeys.length);
-            if (configKeys.length <= 1 || configKeys.indexOf('url') === -1) {
-                this.configValid = false;
-                this.alert.showAlert(constants.idmMessage, constants.featureNotConfigured, 
-                  constants.alertButtons, constants.showDocs, constants.idmUrl);
-            } else {
-                this.configValid = true;
-            }
         }
     }
 }
